aws_chat_app_file/requirements.txt
-e .
aws.cdk.aws_ec2 
aws.cdk.aws_ecs
aws.cdk.aws_ecs_patterns
